pmu: remove event_download_local workaround script
intel_pmu needs to download a config file based on the
CPU model.  When generating VNF images we don't have
access to the actual vCPU that will be used, so we
can't predownload.  This code was meant to be a fix
for that by downloading all the configs and then
selecting one that matched the vCPU.

However we have license issues with intel_pmu event_download GPLv2 code,
so remove it for now.

Change-Id: I5257ff7c4ddc1d40537dadb29efa40d1d68cb852
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>