Replace subprocess "check_output" with "Popen"
"check_output" is a blocking wrapper for "Popen" which returns the output
of the command execution or raises an exception in case of error.

"Popen" is a non-blocking function that allows to create asynchronous
tasks. It returns any possible execution error but doesn't raise an
exception; this is delegated to the developer.

This code is used in the Yardstick CLI base test class.

Change-Id: Ie3e1228b2d40cb306354447653678bf581bc9697
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>