Update patch set 25
Patch Set 25:

i downloaded the current patch set and tox gives me an error for pylint, though according to jekins this patch set is a 0.05 of 10.0 improvement over the parent commit and most of the problems are missing class and method comments.  the important parts to me are:

 ************* Module tests.unit.benchmark.contexts.test_heat
 C:124, 0: Line too long (135/99) (line-too-long)
 C:125, 0: Line too long (146/99) (line-too-long)
 ************* Module tests.unit.network_services.vnf_generic.vnf.test_udp_replay
 C: 37, 0: Import "from tests.unit.network_services.vnf_generic.vnf.test_base import mock_ssh" should be placed at the top of the module (wrong-import-position)
 ************* Module tests.unit.network_services.vnf_generic.vnf.test_vfw_vnf
 ************* Module 
yardstick.benchmark.scenarios.networking.vnf_generic
 C:425, 0: Line too long (100/99) (line-too-long)

and the unit test ones are not as important as the vnf_generic one.

jenkins seems to think the coverage is insufficient, but tox didn't tell me that when i ran the tests.

Patch-set: 25
Reviewer: Gerrit User 6831 <6831@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0