Update patch set 3
Patch Set 3: Code-Review+1

Practically speaking, retrospectively adding docstrings to describe all the methods in a file once you make a change makes it harder to contribute.

We should encourage the use of docstrings to accompany new testcases and feature code, but requireing it is impractical for now.

Patch-set: 3
Label: Code-Review=+1