Update patch set 1
Patch Set 1:

Is there a Jira ticket to go with this?

Nit: there are some style fixes included in this change as well.
I know that these are mandatory in order to pass the gate, however, perhaps it would be a good idea to out the rename from the style fixes.

Patch-set: 1