Update patch set 1
Patch Set 1:

IMO this is a minor fix, and needed for the following patch

"I know that these are mandatory in order to pass the gate, however, perhaps it would be a good idea to out the rename from the style fixes" No, this something exceptional. I don't think we should make this exception.

Patch-set: 1