Add get function in ssh.py
JIRA: YARDSTICK-904

This task is about to implement a get function in ssh.py to
fetch desired files/folder in a remote VM to a local path.

In some test cases, test tools will produce results in files.
We want retain the original result files and archive them in
Yardstick docker for future usage.

Change-Id: Ifd333eb044ce31cffa6f5b2a8f6a46648a8858d1
Signed-off-by: JingLu5 <lvjing5@huawei.com>