Update patch set 1
Patch Set 1: Code-Review-1

(8 comments)

This patch is the first one of many to deploy an infraestructure with several VMs and then deploy OpenStack and Yardstick on them.

We should have a common ansible script and inside it, calls to other scripts:

./ansible/infra_deploy.yml
  -- > ./ansible/roles/infra_check_requirements

Please, add a Jira task under this epic https://jira.opnfv.org/browse/YARDSTICK-938, and make the reference in the commit message.

Patch-set: 1
Label: Code-Review=-1