NSB: Move password from context to scenario
Password specified in the 'contexts' section of an OpenStack test case
(samples/vnf_samples/nsut/router/tc_heat_external_* ) doesn't work.
The variable 'password' is inaccessible in the instantiate function of
RouterVNF class.

With this change, the user will only need to specify the VNF password in the
scenario section for OpenStack test cases.

Other changes:
* Replaced sleep with wait()
* Revert pass to password
* Change RAM value to 4096 to be consistent with SRIOV and OVS test cases
* Set iterations to 10
* Enable IPv4 forwarding in VNF

Change-Id: I7126b7dbc5c20870200d526cfaae2fe73bc12d56
Signed-off-by: Dino Simeon Madarang <dinox.madarang@intel.com>
Reviewed-by: Dhaval Patel <dhaval.r.patel@intel.com>
Reviewed-by: Ross Brattain <ross.b.brattain@intel.com>
Reviewed-by: Alain Jebara <alain.jebara@intel.com>
Reviewed-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>
Reviewed-by: Edward MacGillivray <edward.s.macgillivray@intel.com>
Reviewed-by: Emma Foley <emma.l.foley@intel.com>