Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

You are moving the password definition from context to scenario, not removing the password.

To access to the machine to set the network config, you need this password or the ssh keys.

This patch is doing nothing but changing the position of a variable.

Patch-set: 1
Label: Code-Review=-1