Add arguments to the traffic profile render
In order to render configurable traffic profiles in NSB test cases,
a new variable is introduced: "extra_arg". The content of
this variable is added to the VNFD render data, under a key called
"extra_args".

This will allow the user to define Jinja templates for traffic profiles.

E.g.:

  $ cat test_case_definition.yml
  scenarios:
  - type: NSPerf
    traffic_profile: traffic_profile.yml
    extra_args:
      vports: 10

  $ cat traffic_profile.yml
  {% set vports = get(extra_args, 'vports', '0') or 4 %}
  {% for vport in range(vports|int) %}
  uplink_{{vport}}:
    data...
  {% endfor %}

JIRA: YARDSTICK-946

Change-Id: Ib3c1f2d89efa012772edf2156e97d5f4742a6b80
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>