[WIP] Make 'Scenario' classes plugable
This is a POC to make yardstick.benchmark.scenario.base:Scenario classes
plugable.

A new entry point is added to the setup. This entry point could be extended
in other plugin projects to add new Scenario classes.

Please, take a look at [1]. This is a Yardstick plugin example project.
Clone the project and execute, from the project directory:

  $ sudo -EH python setup.py install

This will add a new module to python ('yardstick-new-plugin') and a
new Scenario class.

Now list the scenarios in Yardstick:

  $ yardstick scenario list
    ...
    | SpecCPU2006_for_VM    | Spec CPU2006 benchmark for Virtual Machine |
    | SpecCPU2006           | Spec CPU2006 benchmark                     |
    | Dummy2                | Execute Dummy (v2!) echo                   |
    +-----------------------+--------------------------------------------+

[1] https://github.com/ralonsoh/yardstick_new_plugin

Change-Id: Ib70ee9bf4dc7ff91d1dd6377317b313288e36bff
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>