Update patch set 2
Patch Set 2:

(3 comments)

Thanks for reviewing the code Rodolfo.

You are moving the password definition from context to scenario, not removing the password.
DM: You are right. I've changed the subject for this.

To access to the machine to set the network config, you need this password or the ssh keys.
DM: Username is set in context and password in scenario. This is for routervnf OpenStack test cases only.

This patch is doing nothing but changing the position of a variable.
DM: The current code has password in context that doesn't work and another in router_vnf.py which requires
users to modify python script and call 'python setup install.py'.

Patch-set: 2
Label: Verified=0