Update patch set 1
Patch Set 1: Code-Review-1

This is a program needed in the CI tasks, not a library needed for execution or testing. So this module shouldn't be installed from requirements.txt or test-requirements.txt. This module should be installed:
- Container
  $cat docker/Dockerfile
  ...
  RUN pip install appdirs==1.4.0 pyopenssl==17.5.0 python-openstackclient==3.11.0

- Baremetal installation:
  This program is not used, because the bash script which needs the OpenStack client, is only executed along with the container installation. Furthermore, this bash script has been replaced with an ansible script, which doesn't use the client.

To make this work (for now), this program should be installed in the container image.

But the definitive solution is to switch, in the CI, to the ansible scripts and get rid of the bash scripts present in the repo.

Patch-set: 1
Label: Code-Review=-1