Update patch set 1
Patch Set 1: Code-Review+2

This is a contradiction: we are installing all the needed packages for baremetal execution (Yardstick on baremetal) and then we are downloading and executing the docker container (which already has all those packages).

But, if anyone has Yardstick installed on the host and needs to install the NSB part (DPDK, Trex, etc), this is needed.

I've open https://jira.opnfv.org/browse/YARDSTICK-1002 in order to start working on the installer improvements.

Patch-set: 1
Reviewer: Gerrit User 7075 <7075@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2