Update patch set 1
Patch Set 1:

Also, another thing to notice is, even though we are installing yardstick in docker but we are mapping "/opt" patch in docker volume with host(where the sample vnf binaries are built). 
Obviously because to run BM tests we need to need to dpdk and sample-vnf build on host.

Patch-set: 1