Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

We need to reduce the number of sample testcases.

This is almost the same as "tc_heat_rfc2544_ipv4_1rule_1flow_64B_trex.yaml", but adding the provider network information.

Instead of this, the mentioned file should be modified in order to make it configurable.

For example:

---
{% set provider = provider|default(None) %}
{% set physical_network = physical_network|default(None) %}

...

{% if provider and physical_network %}
      provider: {{ provider }}
      physical_network: {{ physical_network }}
{% endif %}

Patch-set: 1
Label: Code-Review=-1