Update patch set 2
Patch Set 2:

Hi guys,
Thanks for the advice, I agree using 'os-release' can be a better way to determine the OS distribution.
But as HA testcases are used by dovetail on various platforms, we need to test this method more carefully on more SUTs.
Since there is not much time left in fraser release, I think we should create a jira for this task and implement it in the next release.
And this patch will only focus on the documentation refine base on current HA test method.

Patch-set: 2