Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > Hi Dimitrios,
 > Can you add a Jira ticket to yardstick for this?
 > 
 > Please include references to fuel for this particular scenario, as
 > well as the scenario description itself.
 > Is there really a requirement for 7 nodes when ODL is set-up? it
 > seems like overkill.

Hello Emma.

Yes I will create a jira ticket no problem.

I am not sure that I understood what you mean about the references to fuel and the scenario description. Do you want from me to create also Jira Tickets in Fuel project ? Do you want from me to add a reference for the ODL testcase in the commit message ?

Can you please exaplain to me ? Thanks!

Regarding the number of nodes.Those nodes are all the nodes that the Fuel installer creates when it is deployed. Maybe some of them are not necessary for the Yardstick testcases for now (node 7) but maybe they are necessary into the the future SO I thought it will be better to include all of them so we do not need to update the pod.yaml into the future again.

Patch-set: 4