Update patch set 5
Patch Set 5:

> I think creating another pod file(like pod_odl.yaml) for odl
 > deployment scenario might be better

Hello Kanglin,

I am not sure that this is a good practice because in Fuel ODL is deployed as a standalone not in a clustering mode in noHA and HA deployments.
So tha means that we are gonna have a pod_odl.yaml file for only one node the ODL node. In addition to that in the future we should maintain two files instead of one.

Patch-set: 5
Label: Verified=0