Make Sample VNF hugepages size configurable
The amount of hugepages claimed for a Sample VNF is always 16GB. This
value is excesive for most of the Sample VNF applications (except for vPE).

Making this parameter configurable we allow to spawn smaller VMs by using
less hugepages (in case of StandAlone and OpenStack deployments).

Because this parameter depends on the Scenario and the type of VNF executed,
the parameter is located in:

  scenarios: {options: hugepages_gb} # number of GB of hugepages claimed

PENDING: document this new parameter. A new userguide section should be
         created to document all "scenario" sections and parameters.

JIRA: YARDSTICK-1061

Change-Id: I6f082e105289bd01781be18f2fecbe0ba2fdfdee
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>