[userguide] Use literalinclude directive in 05-operation
Uses literalinclude directive instead of having a copy of the file content in documentation.
This means that any update to this file is detected, and the most up-to-date content is used.

If a testcase is updated, the updates wil be picked up.
The examples used are ones that are tested reqularly.

JIRA: YARDSTICK-1066

Change-Id: I15d967eb04d1e6f4348683d8746fa9989dceda94
Signed-off-by: Emma Foley <emma.l.foley@intel.com>
(cherry picked from commit ee9995c42368ee04ba06cd101b086c57dbbb3bba)