Install RabitMQ for RPC messaging between processes
In NetworkServices Tescases, the TGs (traffic generators) run the traffic in
a separate process. In order to synchronize the traffic injection and the
runner interval loops, an RPC server is needed to publish/subscribe events.

RabbitMQ is a well supported MQ in Linux (used in OpenStack or collectd)
and supported by Python implemented projects like oslo.messaging [1].

RabbitMQ default configuration:
- Port: 5672
- User/password: yardstick/yardstick

[1]https://github.com/openstack/oslo.messaging

JIRA: YARDSTICK-1068

Change-Id: I15db294ee430fb38e574a59b9ce1bf0f8b651a67
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>