Change machine architecture retrieval method
Changed the method to retrieve the architecture of the platform.
"uname -p" returns an string defining this architecture. The set
of possible outputs are described in [1].

In Yardstick CI two architectures are used:
- COMPUTE_ARCH = aarch64 --> YARD_IMG_ARCH = arm64
- COMPUTE_ARCH = x86_64  --> YARD_IMG_ARCH = amd64

"YARD_IMG_ARCH" variable is added as input parameter to
"build_yardstick_image.yml" ansible script, in order to generate
the correct image for the running architecture.

[1]https://stackoverflow.com/questions/45125516/possible-values-for-uname-m

JIRA: YARDSTICK-1058

Change-Id: I33954c236695478eaae34fcfca2e9d5b40791bbf
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>