yardstick image: use growpart instead of parted
parted doesn't handle growing GPT partition correctly, so switch to
growpart

see https://github.com/opnfv/fuel/blob/master/mcp/scripts/lib.sh#L133

for example

If the yardstick image uses GPT then parted complains about the GPT
table

The warning is "Warning: Not all of the space available to
DEVICE appears to be used, you can fix the GPT to use all
of the space (an extra 500 blocks) or continue with the
current setting?"

The backup GPT table is not at the end of the
disk, as it should be. This might mean that
another operating system believes the disk is
smaller. Fix, by moving the backup to the end
(and removing the old backup)?

echo 'f' to fix the GPT table

Instead use growpart

JIRA: YARDSTICK-1058
JIRA: FUEL-359

Change-Id: Ib0e90feef1e3766408b4f742036f2a48af942f32
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>