Fix retrieving "options" section in "scenario"
In [1] a new method to rename the "scenario" host names was implemented.
This method parses all sections with host names and applies the
"qualified" name.

Some malformed test cases define the "options" section without content.
This causes that [2] retrieves "None" value instead of an empty
dictionary. This possibility must be handled in the new method.

[1]I44da30dac562c1a4166e084645ae91c17798651d
[2]https://github.com/opnfv/yardstick/blob/4b8b674b65830a24230faed71e8d9a1048139c89/yardstick/benchmark/core/task.py#L630

JIRA: YARDSTICK-1073

Change-Id: I8864b428734ead8c5aa39de5091d3a2a691be060
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>