restore Heat stack failure logs for CI
We need to see what Heat failures occured
so that we can debug CI failures.

The only way to do this seems to be to use the
private event_utils.get_events function to
query for all failures

Also since we had a failure go ahead and dump the
Heat template that failed.

Change-Id: Ia8f4e5ba7e280fb9086519680d5ee90a2b442e6b
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>