task: don't hide exceptions in Task
If the Task raised an exception we currently hide it
and replace it with RuntimeError.  This is bad.

If an exception occured, then we don't have a result so
re-raise the original exception.

Or we could log the traceback and raise RuntimeError, but
that doesn't seem to be a good idea.

Change-Id: I7e6fa41fc1d36f6d438a1602ab60cb41ffbee1e9
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>