Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

But my point with this bug was to explain what we need to install in the baremetal hosts to run a SampleVNF.

We have an ansible script, ./ansible/ubuntu_server_baremetal_deploy_samplevnfs.yml, to make this deployment but is not documented. How do we run this script? Is there any configuration to be done?

Patch-set: 2
Label: Code-Review=-1