NSB NFVi PROX VPE losing many packets
JIRA: YARDSTICK-1105

Running a vPE test ends up with up to 25% drooped packets. This is due
to tolerated loss being incorrectly configured, test durationwas too short
and test interval was too short.

Note This is due to the fact that vPE use case have by default a
tolerated loss of 100%, Now changed to 1%.

Change-Id: I93e1560c80c2ff2465efba4ad916322dd5312b7b
Signed-off-by: Daniel Martin Buckley <daniel.m.buckley@intel.com>