Update patch set 1
Patch Set 1:

The `placement groups` controls policy of the VM placement but doesn't control which compute host is using for the specific VM (VNF or TG). It has nothing related to OpenStack host aggregates and availability zone, that's why a new option `availability_zone` has been implemented for for this task.

Patch-set: 1
Label: Verified=0