[WIP] Add new runner "iteration_ipc"
This new runner runs a configurable number of times before it returns. Each
iteration has a configurable timeout. The loop control depends on the
feedback received from the running VNFs. The context PIDs from the VNFs
to listen the messages from are given in the scenario "setup" method.

This runner relays on the messaging queue (RabbitMQ) installed in the
jumphost. The messages topic is "VNF", because those are messages sent
from the VNF controlers.

The available methods are listed in common.messsaging (VNF_METHOD_*):
- started: VNF app started
- finisehd: VNF app finished
- iteration: VNF app execution loop started

TODO:
- Unit testing
- Integration tests (SampleVNF test cases)

JIRA: YARDSTICK-1127

Change-Id: I7537b134a4d99434805115984a63fc71414a7d70
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>