disable DPDK images for ARM
We are trying to compile DPDK for the os-nosdn-ovs-ha scenario because we added TC042  "Measure network latency using testpmd and pktgen-dpdk".

However it looks like TC042 is only enabled for compass in opnfv_os-nosdn-kvm_ovs_dpdk-ha

https://gerrit.opnfv.org/gerrit/#/c/54659/1/tests/opnfv/test_suites/opnfv_os-nosdn-kvm_ovs_dpdk-ha_daily.yaml

-
  file_name: opnfv_yardstick_tc042.yaml
  constraint:
      installer: compass
      pod: huawei-pod1

We don't actually need TC042 and DPDK or os-nosdn-ovs-ha.

But in order to run TC042 we have to know beforehand that we are going to use DPDK, so we have a conditional in the load_images.sh script that tries to guess when we will need a DPDK image.

The conditional only checks for *[_-]ovs[_-]*
if [[ $DEPLOY_SCENARIO == *[_-]ovs[_-]* ]]; then
    ansible-playbook \
             -e img_property="nsb" \
             -e YARD_IMG_ARCH=${YARD_IMG_ARCH} \
             -vvv -i inventory.ini build_yardstick_image.yml
    if [ ! -f "${QCOW_NSB_IMAGE}" ]; then
        echo "Failed building QCOW NSB image"
        exit 1
    fi
fi

For this release we won't be able to get DPDK images to compile on ARM, we should fix this conditional and not try to compile.

Also change the shell glob to match *[_-]ovs_dpdk[_-]* to only enable DPDK
images for ovs_dpdk scenarios

JIRA: YARDSTICK-1124

Change-Id: I03142e6b8aa00dc9cdf9b446868fb723cd10198b
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>