Update patch set 1
Patch Set 1:

as mentioned in YARDSTICK-1124, we are only using DPDK for one testcase and only on compass pod.

This patch maybe fixes the dpdk_make_arch issues but it doesn't ensure correct DPDK builds on ARM.  We can't get DPDK fully enabled on ARM for this release, and will disable building DPDK images for tc042 with the other patch.

I'll going to merge this for now and we can continue work next release.

Patch-set: 1