Update patch set 16
Patch Set 16:

Because this script is not used yet, we can merge it and then use the CI we are going to build to test it.

It could be nice to replace the bash installation script with this one.

Next steps:
- Create the VM images (normal and nsb)
- If OpenStack --> upload the images
- Deploy, if needed, the baremetal and standalone machines

All those scripts are already implemented. Is just a matter of adding them to this one.

Patch-set: 16