Add "os_cloud_config" as a new context flag parameter
This new parameter will contain the OpenStack cloud specific configuration
used by Shade client. This new flag is used only in Heat context.

By default, this new parameter (dict) will contain this content:
  'os_cloud_config': {'verify': False}

This field will be used by HeatStack [1] to create a Shade cloud. Shade
retrieves, if not defined, the OpenStack configuration from
"os_client_config". This configuration is used to generate the cloud
configuration, which is the description of the OpenStackCloud returned.

The default parameter defined, "verify", refers to the related bug. By
default, in case of using SSL certificate it will not be verified.

[1] https://github.com/opnfv/yardstick/blob/b338d3091bb0beb89d4ad9f7c144f43a31a19a74/yardstick/orchestrator/heat.py#L47

JIRA: YARDSTICK-1139

Change-Id: I875a7018401b84e51dab775b8194174645d27e06
Signed-off-by: Rodolfo Alonso Hernandez <rodolfo.alonso.hernandez@intel.com>