Update patch set 3
Patch Set 3:

> Yes, sorry for that. Should I modify the Jira ticket and my commit
 > so that it only concerns fixing typos of log or error messages?
Yes please, the commit message and the jira description should be similar.

 > Should I make a separate ticket now for the other changes, or wait
 > for a code quality epic/theme to come out?

Create the ticket, and add the jira reference to [1] under the appropriate heading so we can add it in later.

 > About the tests, could I make a separate Jira ticket for moving
 > them to unit/functional tests?
Yes, absolutely, you can log it as a bug or as an improvement, and list the offending files, if possible, so we can track whether they have all been fixed. You can then add the Jira ticket to [1] under an appropriate theme, if the epic doesn't already exist.

[1] https://etherpad.opnfv.org/p/yardstick_release_g

Patch-set: 3