Update patch set 1
Patch Set 1:

We need to delete INFO because it was replaced by INFO.yaml.

I'm trying to do a git trick to see if I can link the history of INFO to INFO.yaml, so we have a record of past committers.

I thought git would do some rename magic and detect that the files are the same, but it doesn't seem to work.

Patch-set: 1
Reviewer: Gerrit User 1402 <1402@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0